move SidechainParams to new file
